Refactor Project Component
