+++ src/renderer/components/Project/CreatingScenarioModal.js
@@ -1,9 +1,19 @@
 import React from 'react';
-import { Modal } from 'antd';
+import { Modal, Button, Icon } from 'antd';
+import { AsyncError } from '../../utils';
 
-const CreatingScenarioModal = ({ visible, setVisible }) => {
+const CreatingScenarioModal = ({
+  visible,
+  setVisible,
+  error,
+  createScenario
+}) => {
   const hideModal = () => {
     setVisible(false);
+  };
+
+  const retry = () => {
+    createScenario();
   };
 
   return (
@@ -15,7 +25,20 @@ const CreatingScenarioModal = ({ visible, setVisible }) => {
       closable={false}
       maskClosable={false}
     >
-      Creating Scenario
+      {error ? (
+        <div>
+          <AsyncError error={error} />
+          <div style={{ display: 'flex', justifyContent: 'flex-end' }}>
+            <Button onClick={hideModal}>Cancel</Button>
+            <Button onClick={retry}>Retry</Button>
+          </div>
+        </div>
+      ) : (
+        <div>
+          <Icon type="loading" style={{ color: 'blue', margin: 5 }} />
+          <span>Creating Scenario</span>
+        </div>
+      )}
     </Modal>
   );
 };
